@@ -49,7 +49,7 @@ namespace DotNetty.Common.Concurrency
             WakeupTask = new NoOpRunnable();
         }
 
-        protected internal readonly IPriorityQueue<IScheduledRunnable> ScheduledTaskQueue;
+        internal readonly IPriorityQueue<IScheduledRunnable> _scheduledTaskQueue;
         private long _nextTaskId;
 
         protected AbstractScheduledEventExecutor()
@@ -60,8 +60,11 @@ namespace DotNetty.Common.Concurrency
         protected AbstractScheduledEventExecutor(IEventExecutorGroup parent)
             : base(parent)
         {
-            ScheduledTaskQueue = new DefaultPriorityQueue<IScheduledRunnable>();
+            _scheduledTaskQueue = new DefaultPriorityQueue<IScheduledRunnable>();
         }
+
+        /// <summary>TBD</summary>
+        protected abstract bool HasTasks { get; }
 
         [MethodImpl(InlineMethod.AggressiveOptimization)]
         protected static PreciseTimeSpan GetNanos() => PreciseTimeSpan.FromStart;
@@ -89,6 +92,7 @@ namespace DotNetty.Common.Concurrency
             return taskQueue is null || 0u >= (uint)taskQueue.Count;
         }
 
+        [Obsolete("Please use IPriorityQueue{T}.IsEmpty instead.")]
         [MethodImpl(InlineMethod.AggressiveOptimization)]
         protected static bool IsEmpty(IPriorityQueue<IScheduledRunnable> taskQueue)
         {
@@ -103,8 +107,8 @@ namespace DotNetty.Common.Concurrency
         {
             Debug.Assert(InEventLoop);
 
-            IPriorityQueue<IScheduledRunnable> scheduledTaskQueue = ScheduledTaskQueue;
-            if (IsEmpty(scheduledTaskQueue)) { return; }
+            IPriorityQueue<IScheduledRunnable> scheduledTaskQueue = _scheduledTaskQueue;
+            if (scheduledTaskQueue.IsEmpty) { return; }
 
             IScheduledRunnable[] tasks = scheduledTaskQueue.ToArray();
             for (int i = 0; i < tasks.Length; i++)
@@ -112,7 +116,7 @@ namespace DotNetty.Common.Concurrency
                 _ = tasks[i].CancelWithoutRemove();
             }
 
-            ScheduledTaskQueue.ClearIgnoringIndexes();
+            _scheduledTaskQueue.ClearIgnoringIndexes();
         }
 
         internal protected IScheduledRunnable PollScheduledTask() => PollScheduledTask(NanoTime());
@@ -126,10 +130,10 @@ namespace DotNetty.Common.Concurrency
         {
             Debug.Assert(InEventLoop);
 
-            if (ScheduledTaskQueue.TryPeek(out IScheduledRunnable scheduledTask) &&
+            if (_scheduledTaskQueue.TryPeek(out IScheduledRunnable scheduledTask) &&
                 scheduledTask.DeadlineNanos <= nanoTime)
             {
-                _ = ScheduledTaskQueue.TryDequeue(out _);
+                _ = _scheduledTaskQueue.TryDequeue(out _);
                 scheduledTask.SetConsumed();
                 return scheduledTask;
             }
@@ -142,7 +146,7 @@ namespace DotNetty.Common.Concurrency
         /// </summary>
         protected long NextScheduledTaskNanos()
         {
-            if (ScheduledTaskQueue.TryPeek(out IScheduledRunnable nextScheduledRunnable))
+            if (_scheduledTaskQueue.TryPeek(out IScheduledRunnable nextScheduledRunnable))
             {
                 return nextScheduledRunnable.DelayNanos;
             }
@@ -155,7 +159,7 @@ namespace DotNetty.Common.Concurrency
         /// </summary>
         protected long NextScheduledTaskDeadlineNanos()
         {
-            if (ScheduledTaskQueue.TryPeek(out IScheduledRunnable nextScheduledRunnable))
+            if (_scheduledTaskQueue.TryPeek(out IScheduledRunnable nextScheduledRunnable))
             {
                 return nextScheduledRunnable.DeadlineNanos;
             }
@@ -165,9 +169,12 @@ namespace DotNetty.Common.Concurrency
         [MethodImpl(InlineMethod.AggressiveOptimization)]
         protected IScheduledRunnable PeekScheduledTask()
         {
-            //IPriorityQueue<IScheduledRunnable> scheduledTaskQueue = ScheduledTaskQueue;
-            //return !IsNullOrEmpty(scheduledTaskQueue) && scheduledTaskQueue.TryPeek(out IScheduledRunnable task) ? task : null;
-            return ScheduledTaskQueue.TryPeek(out IScheduledRunnable task) ? task : null;
+            return _scheduledTaskQueue.TryPeek(out IScheduledRunnable task) ? task : null;
+        }
+
+        protected bool TryPeekScheduledTask(out IScheduledRunnable task)
+        {
+            return _scheduledTaskQueue.TryPeek(out task);
         }
 
         /// <summary>
@@ -175,7 +182,7 @@ namespace DotNetty.Common.Concurrency
         /// </summary>
         protected bool HasScheduledTasks()
         {
-            return ScheduledTaskQueue.TryPeek(out IScheduledRunnable scheduledTask) && scheduledTask.DeadlineNanos <= PreciseTime.NanoTime();
+            return _scheduledTaskQueue.TryPeek(out IScheduledRunnable scheduledTask) && scheduledTask.DeadlineNanos <= PreciseTime.NanoTime();
         }
 
         public override IScheduledTask Schedule(IRunnable action, TimeSpan delay)
@@ -486,7 +493,19 @@ namespace DotNetty.Common.Concurrency
         internal void ScheduleFromEventLoop(IScheduledRunnable task)
         {
             // nextTaskId a long and so there is no chance it will overflow back to 0
-            _ = ScheduledTaskQueue.TryEnqueue(task.SetId(++_nextTaskId));
+            var nextTaskId = ++_nextTaskId;
+            if (nextTaskId == long.MaxValue) { _nextTaskId = 0; }
+
+            var isBacklogEmpty = !HasTasks && _scheduledTaskQueue.IsEmpty;
+
+            _ = _scheduledTaskQueue.TryEnqueue(task.SetId(nextTaskId));
+
+            if (isBacklogEmpty)
+            {
+                // 在 Libuv.LoopExecutor 中，当任务执行完毕，清空任务队列后，后续如果只有 ScheduledTask 入队的情况下，
+                // 并不会激发线程进行任务处理，需唤醒
+                EnusreWakingUp(true);
+            }
         }
 
         private IScheduledRunnable Schedule(IScheduledRunnable task)
@@ -520,7 +539,7 @@ namespace DotNetty.Common.Concurrency
         {
             if (InEventLoop)
             {
-                _ = ScheduledTaskQueue.TryRemove(task);
+                _ = _scheduledTaskQueue.TryRemove(task);
             }
             else
             {
@@ -555,6 +574,10 @@ namespace DotNetty.Common.Concurrency
         {
             return true;
         }
+
+        /// <summary>TBD</summary>
+        /// <param name="inEventLoop"></param>
+        protected virtual void EnusreWakingUp(bool inEventLoop) { }
 
         sealed class NoOpRunnable : IRunnable
         {
