@@ -135,6 +135,9 @@ namespace DotNetty.Common.Concurrency
         /// </summary>
         public int BacklogLength => _taskQueue.Count;
 
+        /// <inheritdoc />
+        protected override bool HasTasks => _taskQueue.NonEmpty;
+
         void Loop(object s)
         {
             SetCurrentExecutor(this);
@@ -230,7 +233,7 @@ namespace DotNetty.Common.Concurrency
 
         protected override IEnumerable<IEventExecutor> GetItems() => new[] { this };
 
-        protected void WakeUp(bool inEventLoop)
+        protected internal virtual void WakeUp(bool inEventLoop)
         {
             if (!inEventLoop || (Volatile.Read(ref v_executionState) == ST_SHUTTING_DOWN))
             {
@@ -613,7 +616,7 @@ namespace DotNetty.Common.Concurrency
 
         private bool FetchFromScheduledTaskQueue()
         {
-            if (ScheduledTaskQueue.IsEmpty) { return true; }
+            if (_scheduledTaskQueue.IsEmpty) { return true; }
 
             var nanoTime = PreciseTime.NanoTime();
             IScheduledRunnable scheduledTask = PollScheduledTask(nanoTime);
@@ -622,7 +625,7 @@ namespace DotNetty.Common.Concurrency
                 if (!_taskQueue.TryEnqueue(scheduledTask))
                 {
                     // No space left in the task queue add it back to the scheduledTaskQueue so we pick it up again.
-                    _ = ScheduledTaskQueue.TryEnqueue(scheduledTask);
+                    _ = _scheduledTaskQueue.TryEnqueue(scheduledTask);
                     return false;
                 }
                 scheduledTask = PollScheduledTask(nanoTime);
@@ -639,7 +642,7 @@ namespace DotNetty.Common.Concurrency
                 _emptyEvent.Reset();
                 if (!_taskQueue.TryDequeue(out task) && !IsShuttingDown) // revisit queue as producer might have put a task in meanwhile
                 {
-                    if (ScheduledTaskQueue.TryPeek(out IScheduledRunnable nextScheduledTask))
+                    if (_scheduledTaskQueue.TryPeek(out IScheduledRunnable nextScheduledTask))
                     {
                         PreciseTimeSpan wakeupTimeout = nextScheduledTask.Deadline - PreciseTimeSpan.FromStart;
                         if (wakeupTimeout.Ticks > 0L) // 此处不要 ulong 转换
