@@ -31,6 +31,7 @@ namespace DotNetty.Handlers.Tls
     using System;
     using System.Diagnostics;
     using System.Net.Security;
+    using System.Runtime.CompilerServices;
     using System.Threading.Tasks;
     using DotNetty.Common.Utilities;
     using DotNetty.Transport.Channels;
@@ -47,99 +48,123 @@ namespace DotNetty.Handlers.Tls
         private bool EnsureAuthenticated(IChannelHandlerContext ctx)
         {
             var oldState = State;
-            if (!oldState.HasAny(TlsHandlerState.AuthenticationStarted))
+            if (oldState.HasAny(TlsHandlerState.AuthenticationStarted))
             {
-                State = oldState | TlsHandlerState.Authenticating;
-                if (_isServer)
-                {
-#if NETCOREAPP_2_0_GREATER || NETSTANDARD_2_0_GREATER
-                    // Adapt to the SslStream signature
-                    ServerCertificateSelectionCallback selector = null;
-                    if (_serverCertificateSelector is object)
-                    {
-                        X509Certificate LocalServerCertificateSelection(object sender, string name)
-                        {
-                            ctx.GetAttribute(SslStreamAttrKey).Set(_sslStream);
-                            return _serverCertificateSelector(ctx, name);
-                        }
-                        selector = new ServerCertificateSelectionCallback(LocalServerCertificateSelection);
-                    }
-
-                    var sslOptions = new SslServerAuthenticationOptions()
-                    {
-                        ServerCertificate = _serverCertificate,
-                        ServerCertificateSelectionCallback = selector,
-                        ClientCertificateRequired = _serverSettings.NegotiateClientCertificate,
-                        EnabledSslProtocols = _serverSettings.EnabledProtocols,
-                        CertificateRevocationCheckMode = _serverSettings.CheckCertificateRevocation ? X509RevocationMode.Online : X509RevocationMode.NoCheck,
-                        ApplicationProtocols = _serverSettings.ApplicationProtocols // ?? new List<SslApplicationProtocol>()
-                    };
-                    if (_hasHttp2Protocol)
-                    {
-                        // https://tools.ietf.org/html/rfc7540#section-9.2.1
-                        sslOptions.AllowRenegotiation = false;
-                    }
-                    _sslStream.AuthenticateAsServerAsync(sslOptions, CancellationToken.None)
-                              .ContinueWith(s_handshakeCompletionCallback, this, TaskContinuationOptions.ExecuteSynchronously);
-#else
-                    _sslStream.AuthenticateAsServerAsync(_serverCertificate,
-                                                         _serverSettings.NegotiateClientCertificate,
-                                                         _serverSettings.EnabledProtocols,
-                                                         _serverSettings.CheckCertificateRevocation)
-                              .ContinueWith(s_handshakeCompletionCallback, this, TaskContinuationOptions.ExecuteSynchronously);
-#endif
-                }
-                else
-                {
-#if NETCOREAPP_2_0_GREATER || NETSTANDARD_2_0_GREATER
-                    LocalCertificateSelectionCallback selector = null;
-                    if (_userCertSelector is object)
-                    {
-                        X509Certificate LocalCertificateSelection(object sender, string targetHost, X509CertificateCollection localCertificates, X509Certificate remoteCertificate, string[] acceptableIssuers)
-                        {
-                            ctx.GetAttribute(SslStreamAttrKey).Set(_sslStream);
-                            return _userCertSelector(ctx, targetHost, localCertificates, remoteCertificate, acceptableIssuers);
-                        }
-                        selector = new LocalCertificateSelectionCallback(LocalCertificateSelection);
-                    }
-                    var sslOptions = new SslClientAuthenticationOptions()
-                    {
-                        TargetHost = _clientSettings.TargetHost,
-                        ClientCertificates = _clientSettings.X509CertificateCollection,
-                        EnabledSslProtocols = _clientSettings.EnabledProtocols,
-                        CertificateRevocationCheckMode = _clientSettings.CheckCertificateRevocation ? X509RevocationMode.Online : X509RevocationMode.NoCheck,
-                        LocalCertificateSelectionCallback = selector,
-                        ApplicationProtocols = _clientSettings.ApplicationProtocols
-                    };
-                    if (_hasHttp2Protocol)
-                    {
-                        // https://tools.ietf.org/html/rfc7540#section-9.2.1
-                        sslOptions.AllowRenegotiation = false;
-                    }
-                    _sslStream.AuthenticateAsClientAsync(sslOptions, CancellationToken.None)
-                              .ContinueWith(s_handshakeCompletionCallback, this, TaskContinuationOptions.ExecuteSynchronously);
-#else
-                    _sslStream.AuthenticateAsClientAsync(_clientSettings.TargetHost,
-                                                         _clientSettings.X509CertificateCollection,
-                                                         _clientSettings.EnabledProtocols,
-                                                         _clientSettings.CheckCertificateRevocation)
-                              .ContinueWith(s_handshakeCompletionCallback, this, TaskContinuationOptions.ExecuteSynchronously);
-#endif
-                }
-                return false;
+                return oldState.Has(TlsHandlerState.Authenticated);
             }
 
-            return oldState.Has(TlsHandlerState.Authenticated);
+            State = oldState | TlsHandlerState.Authenticating;
+            BeginHandshake(ctx);
+            return false;
+        }
+
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private void BeginHandshake(IChannelHandlerContext ctx)
+        {
+            if (_isServer)
+            {
+#if NETCOREAPP_2_0_GREATER || NETSTANDARD_2_0_GREATER
+                // Adapt to the SslStream signature
+                ServerCertificateSelectionCallback selector = null;
+                if (_serverCertificateSelector is object)
+                {
+                    X509Certificate LocalServerCertificateSelection(object sender, string name)
+                    {
+                        ctx.GetAttribute(SslStreamAttrKey).Set(_sslStream);
+                        return _serverCertificateSelector(ctx, name);
+                    }
+                    selector = new ServerCertificateSelectionCallback(LocalServerCertificateSelection);
+                }
+
+                var sslOptions = new SslServerAuthenticationOptions()
+                {
+                    ServerCertificate = _serverCertificate,
+                    ServerCertificateSelectionCallback = selector,
+                    ClientCertificateRequired = _serverSettings.NegotiateClientCertificate,
+                    EnabledSslProtocols = _serverSettings.EnabledProtocols,
+                    CertificateRevocationCheckMode = _serverSettings.CheckCertificateRevocation ? X509RevocationMode.Online : X509RevocationMode.NoCheck,
+                    ApplicationProtocols = _serverSettings.ApplicationProtocols // ?? new List<SslApplicationProtocol>()
+                };
+                _serverSettings.OnAuthenticate?.Invoke(ctx, _serverSettings, sslOptions);
+
+                var cts = new CancellationTokenSource(_serverSettings.HandshakeTimeout);
+                _sslStream.AuthenticateAsServerAsync(sslOptions, cts.Token)
+                          .ContinueWith(
+#if NET
+                                static
+#endif
+                                (t, s) => HandshakeCompletionCallback(t, s), (this, cts), TaskContinuationOptions.ExecuteSynchronously);
+#else
+                _sslStream.AuthenticateAsServerAsync(_serverCertificate,
+                                                     _serverSettings.NegotiateClientCertificate,
+                                                     _serverSettings.EnabledProtocols,
+                                                     _serverSettings.CheckCertificateRevocation)
+                          .ContinueWith((t, s) => HandshakeCompletionCallback(t, s), this, TaskContinuationOptions.ExecuteSynchronously);
+#endif
+            }
+            else
+            {
+#if NETCOREAPP_2_0_GREATER || NETSTANDARD_2_0_GREATER
+                LocalCertificateSelectionCallback selector = null;
+                if (_userCertSelector is object)
+                {
+                    X509Certificate LocalCertificateSelection(object sender, string targetHost, X509CertificateCollection localCertificates, X509Certificate remoteCertificate, string[] acceptableIssuers)
+                    {
+                        ctx.GetAttribute(SslStreamAttrKey).Set(_sslStream);
+                        return _userCertSelector(ctx, targetHost, localCertificates, remoteCertificate, acceptableIssuers);
+                    }
+                    selector = new LocalCertificateSelectionCallback(LocalCertificateSelection);
+                }
+                var sslOptions = new SslClientAuthenticationOptions()
+                {
+                    TargetHost = _clientSettings.TargetHost,
+                    ClientCertificates = _clientSettings.X509CertificateCollection,
+                    EnabledSslProtocols = _clientSettings.EnabledProtocols,
+                    CertificateRevocationCheckMode = _clientSettings.CheckCertificateRevocation ? X509RevocationMode.Online : X509RevocationMode.NoCheck,
+                    LocalCertificateSelectionCallback = selector,
+                    ApplicationProtocols = _clientSettings.ApplicationProtocols
+                };
+                _clientSettings.OnAuthenticate?.Invoke(ctx, _clientSettings, sslOptions);
+
+                var cts = new CancellationTokenSource(_clientSettings.HandshakeTimeout);
+                _sslStream.AuthenticateAsClientAsync(sslOptions, cts.Token)
+                          .ContinueWith(
+#if NET
+                                static
+#endif
+                                (t, s) => HandshakeCompletionCallback(t, s), (this, cts), TaskContinuationOptions.ExecuteSynchronously);
+#else
+                _sslStream.AuthenticateAsClientAsync(_clientSettings.TargetHost,
+                                                     _clientSettings.X509CertificateCollection,
+                                                     _clientSettings.EnabledProtocols,
+                                                     _clientSettings.CheckCertificateRevocation)
+                          .ContinueWith((t, s) => HandshakeCompletionCallback(t, s), this, TaskContinuationOptions.ExecuteSynchronously);
+#endif
+            }
+        }
+
+        private static void HandshakeCompletionCallback(Task task, object s)
+        {
+#if NETCOREAPP_2_0_GREATER || NETSTANDARD_2_0_GREATER
+            var (self, cts) = ((TlsHandler self, CancellationTokenSource cts))s;
+            cts.Dispose();
+#else
+            var self = (TlsHandler)s;
+#endif
+            var capturedContext = self.CapturedContext;
+            if (capturedContext.Executor.InEventLoop)
+            {
+                HandleHandshakeCompleted(task, self);
+            }
+            else
+            {
+                capturedContext.Executor.Execute(s_handshakeCompletionCallback, task, self);
+            }
         }
 
         private static void HandleHandshakeCompleted(Task task, TlsHandler self)
         {
             var capturedContext = self.CapturedContext;
-            if (!capturedContext.Executor.InEventLoop)
-            {
-                capturedContext.Executor.Execute(s_handshakeCompletionCallback, task, self);
-                return;
-            }
             var oldState = self.State;
 
             if (task.IsSuccess())
