@@ -23,10 +23,10 @@
 namespace DotNetty.Handlers.Tls
 {
     using System;
-    using System.Collections.Generic;
     using System.Diagnostics;
     using System.Runtime.CompilerServices;
     using System.Runtime.ExceptionServices;
+    using System.Net.Security;
     using System.Threading.Tasks;
     using DotNetty.Buffers;
     using DotNetty.Common.Concurrency;
@@ -38,23 +38,48 @@ namespace DotNetty.Handlers.Tls
 
     partial class TlsHandler
     {
-        private Task _lastContextWriteTask;
+        private IPromise _lastContextWritePromise;
+        private volatile int v_wrapDataSize = TlsUtils.MAX_PLAINTEXT_LENGTH;
+
+        /// <summary>
+        /// Gets or Sets the number of bytes to pass to each <see cref="SslStream.Write(byte[], int, int)"/> call.
+        /// </summary>
+        /// <remarks>
+        /// This value will partition data which is passed to write
+        /// <see cref="Write(IChannelHandlerContext, object, IPromise)"/> The partitioning will work as follows:
+        /// <ul>
+        /// <li>If <code>wrapDataSize &lt;= 0</code> then we will write each data chunk as is.</li>
+        /// <li>If <code>wrapDataSize > data size</code> then we will attempt to aggregate multiple data chunks together.</li>
+        /// <li>Else if <code>wrapDataSize &lt;= data size</code> then we will divide the data into chunks of <c>wrapDataSize</c> when writing.</li>
+        /// </ul>
+        /// </remarks>
+        public int WrapDataSize
+        {
+            get => v_wrapDataSize;
+            set => v_wrapDataSize = value;
+        }
 
         public override void Write(IChannelHandlerContext context, object message, IPromise promise)
         {
-            if (message is IByteBuffer buf)
+            if (message is not IByteBuffer buf)
             {
-                if (_pendingUnencryptedWrites is object)
-                {
-                    _pendingUnencryptedWrites.Add(buf, promise);
-                }
-                else
-                {
-                    ReferenceCountUtil.SafeRelease(buf);
-                    _ = promise.TrySetException(NewPendingWritesNullException());
-                }
+                InvalidMessage(message, promise);
                 return;
             }
+            if (_pendingUnencryptedWrites is object)
+            {
+                _pendingUnencryptedWrites.Add(buf, promise);
+            }
+            else
+            {
+                ReferenceCountUtil.SafeRelease(buf);
+                _ = promise.TrySetException(NewPendingWritesNullException());
+            }
+        }
+
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private void InvalidMessage(object message, IPromise promise)
+        {
             ReferenceCountUtil.SafeRelease(message);
             _ = promise.TrySetException(ThrowHelper.GetUnsupportedMessageTypeException(message));
         }
@@ -88,7 +113,7 @@ namespace DotNetty.Handlers.Tls
 
         private void WrapAndFlush(IChannelHandlerContext context)
         {
-            if (_pendingUnencryptedWrites.IsEmpty)
+            if (_pendingUnencryptedWrites.IsEmpty())
             {
                 // It's important to NOT use a voidPromise here as the user
                 // may want to add a ChannelFutureListener to the ChannelPromise later.
@@ -118,47 +143,49 @@ namespace DotNetty.Handlers.Tls
         {
             Debug.Assert(context == CapturedContext);
 
+            IByteBufferAllocator alloc = context.Allocator;
             IByteBuffer buf = null;
             try
             {
+                int wrapDataSize = v_wrapDataSize;
                 // Only continue to loop if the handler was not removed in the meantime.
                 // See https://github.com/netty/netty/issues/5860
                 while (!context.IsRemoved)
                 {
-                    List<object> messages = _pendingUnencryptedWrites.Current;
-                    if (messages is null || 0u >= (uint)messages.Count)
-                    {
-                        break;
-                    }
+                    var promise = context.NewPromise();
+                    buf = wrapDataSize > 0
+                        ? _pendingUnencryptedWrites.Remove(alloc, wrapDataSize, promise)
+                        : _pendingUnencryptedWrites.RemoveFirst(promise);
+                    if (buf is null) { break; }
 
-                    if (1u >= (uint)messages.Count) // messages.Count == 1; messages 最小数量为 1
+                    try
                     {
-                        buf = (IByteBuffer)messages[0];
-                    }
-                    else
-                    {
-                        buf = context.Allocator.Buffer((int)_pendingUnencryptedWrites.CurrentSize);
-                        for (int idx = 0; idx < messages.Count; idx++)
+                        _lastContextWritePromise = promise;
+                        _ = buf.ReadBytes(_sslStream, buf.ReadableBytes); // this leads to FinishWrap being called 0+ times
+
+                        if (buf.IsReadable())
                         {
-                            var buffer = (IByteBuffer)messages[idx];
-                            _ = buffer.ReadBytes(buf, buffer.ReadableBytes);
-                            _ = buffer.Release();
+                            _pendingUnencryptedWrites.AddFirst(buf, promise);
+                            // When we add the buffer/promise pair back we need to be sure we don't complete the promise
+                            // later in finishWrap. We only complete the promise if the buffer is completely consumed.
+                            //promise = null;
+                        }
+                        else
+                        {
+                            buf.Release();
                         }
                     }
-                    _ = buf.ReadBytes(_sslStream, buf.ReadableBytes); // this leads to FinishWrap being called 0+ times
-                    _ = buf.Release();
-                    buf = null;
-
-                    var promise = _pendingUnencryptedWrites.Remove();
-                    Task task = _lastContextWriteTask;
-                    if (task is object)
+                    catch (Exception)
                     {
-                        task.LinkOutcome(promise);
-                        _lastContextWriteTask = null;
+                        buf.Release();
+                        promise.TrySetException(s_sslStreamClosedException);
+                        throw;
                     }
-                    else
+                    finally
                     {
-                        _ = promise.TryComplete();
+                        buf = null;
+                        promise = null;
+                        _lastContextWritePromise = null;
                     }
                 }
             }
@@ -186,7 +213,7 @@ namespace DotNetty.Handlers.Tls
                 output.Advance(bufLen);
             }
 
-            _lastContextWriteTask = capturedContext.WriteAsync(output, promise);
+            _ = capturedContext.WriteAsync(output, promise);
         }
 #endif
 
@@ -204,7 +231,7 @@ namespace DotNetty.Handlers.Tls
                 _ = output.WriteBytes(buffer, offset, count);
             }
 
-            _lastContextWriteTask = capturedContext.WriteAsync(output, promise);
+            _ = capturedContext.WriteAsync(output, promise);
         }
 
 #if NETCOREAPP || NETSTANDARD_2_0_GREATER
@@ -218,6 +245,7 @@ namespace DotNetty.Handlers.Tls
             }
             else
             {
+                // TODO 优化
                 future = capturedContext.WriteAndFlushAsync(Unpooled.WrappedBuffer(buffer.ToArray()), promise);
             }
             this.ReadIfNeeded(capturedContext);
