@@ -29,13 +29,18 @@
 namespace DotNetty.Handlers.Tls
 {
     using System.Security.Authentication;
+#if NETCOREAPP_2_0_GREATER || NETSTANDARD_2_0_GREATER
+    using System;
+    using System.Runtime.CompilerServices;
+    using System.Threading;
+#endif
 
     public abstract class TlsSettings
     {
         protected TlsSettings(SslProtocols enabledProtocols, bool checkCertificateRevocation)
         {
-            this.EnabledProtocols = enabledProtocols;
-            this.CheckCertificateRevocation = checkCertificateRevocation;
+            EnabledProtocols = enabledProtocols;
+            CheckCertificateRevocation = checkCertificateRevocation;
         }
 
         /// <summary>Specifies allowable SSL protocols.</summary>
@@ -43,5 +48,39 @@ namespace DotNetty.Handlers.Tls
 
         /// <summary>Specifies whether the certificate revocation list is checked during authentication.</summary>
         public bool CheckCertificateRevocation { get; }
+
+#if NETCOREAPP_2_0_GREATER || NETSTANDARD_2_0_GREATER
+        private static readonly TimeSpan DefaultHandshakeTimeout = TimeSpan.FromSeconds(10);
+        private static readonly TimeSpan MaximumHandshakeTimeout = TimeSpan.FromMilliseconds(int.MaxValue);
+
+        private TimeSpan _handshakeTimeout = DefaultHandshakeTimeout;
+
+        /// <summary>
+        /// Specifies the maximum amount of time allowed for the TLS/SSL handshake. This must be positive and finite. Defaults to 10 seconds.
+        /// </summary>
+        public TimeSpan HandshakeTimeout
+        {
+            get => _handshakeTimeout;
+            set
+            {
+                if (value <= TimeSpan.Zero && value != Timeout.InfiniteTimeSpan || value > MaximumHandshakeTimeout)
+                {
+                    ThrowArgumentOutOfRangeException();
+                }
+                _handshakeTimeout = value != Timeout.InfiniteTimeSpan ? value : MaximumHandshakeTimeout;
+            }
+        }
+
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private static void ThrowArgumentOutOfRangeException()
+        {
+            throw GetArgumentOutOfRangeException();
+
+            static ArgumentOutOfRangeException GetArgumentOutOfRangeException()
+            {
+                return new ArgumentOutOfRangeException("value", "Value must be a positive TimeSpan.");
+            }
+        }
+#endif
     }
 }
