@@ -32,12 +32,26 @@ namespace DotNetty.Handlers.Tls
     using DotNetty.Buffers;
     using DotNetty.Transport.Channels;
 
+    internal enum TlsFraming
+    {
+        Unknown = 0,    // Initial before any frame is processd.
+        BeforeSSL3,     // SSlv2
+        SinceSSL3,      // SSlv3 & TLS
+        Unified,        // Intermediate on first frame until response is processes.
+        Invalid         // Somthing is wrong.
+    }
+
     /// Utilities for TLS packets.
     static class TlsUtils
     {
-        const int MAX_PLAINTEXT_LENGTH = 16 * 1024; // 2^14
-        const int MAX_COMPRESSED_LENGTH = MAX_PLAINTEXT_LENGTH + 1024;
-        const int MAX_CIPHERTEXT_LENGTH = MAX_COMPRESSED_LENGTH + 1024;
+        /// <summary>
+        /// <a href="https://tools.ietf.org/html/rfc5246#section-6.2">2^14</a> which is the maximum sized plaintext chunk
+        /// allowed by the TLS RFC.
+        /// </summary>
+        public const int MAX_PLAINTEXT_LENGTH = 16 * 1024; // 2^14
+        private const int MAX_COMPRESSED_LENGTH = MAX_PLAINTEXT_LENGTH + 1024;
+        private const int MAX_CIPHERTEXT_LENGTH = MAX_COMPRESSED_LENGTH + 1024;
+        private const int GMSSL_PROTOCOL_VERSION = 0x101;
 
         // Header (5) + Data (2^14) + Compression (1024) + Encryption (1024) + MAC (20) + Padding (256)
         public const int MAX_ENCRYPTED_PACKET_LENGTH = MAX_CIPHERTEXT_LENGTH + 5 + 20 + 256;
@@ -105,11 +119,11 @@ namespace DotNetty.Handlers.Tls
 
             if (tls)
             {
-                // SSLv3 or TLS - Check ProtocolVersion
+                // SSLv3 or TLS or GMSSLv1.0 or GMSSLv1.1 - Check ProtocolVersion
                 int majorVersion = buffer.GetByte(offset + 1);
-                if (majorVersion == 3)
+                if (majorVersion == 3 || buffer.GetShort(offset + 1) == GMSSL_PROTOCOL_VERSION)
                 {
-                    // SSLv3 or TLS
+                    // SSLv3 or TLS or GMSSLv1.0 or GMSSLv1.1
                     packetLength = buffer.GetUnsignedShort(offset + 3) + SSL_RECORD_HEADER_LENGTH;
                     if ((uint)packetLength <= SSL_RECORD_HEADER_LENGTH)
                     {
